+++ bin/pgstatsmon.js
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  *
- * Copyright (c) 2017, Joyent, Inc.
+ * Copyright (c) 2018, Joyent, Inc.
  */
 
 /*
@@ -47,8 +47,9 @@ function main()
 	});
 
 	log.info('config', config);
+
 	config['log'] = log;
-	pgstatsmon(config);
+	pgstatsmon(config).start();
 }
 
 main();