+++ CHANGES.md
@@ -1,6 +1,10 @@
 # pgstatsmon Changelog
 
 ## Not yet released.
+* #8 create a Postgres user for pgstatsmon
+* #7 add 'release' and 'publish' targets
+* #5 add and improve queries
+* #4 improve connection handling and backend discovery
 * #3 need a test suite
 * #2 repo housekeeping
 * #1 support Prometheus-style metric collection